+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Stencil Changelog
 
-## Master
+## 0.14.0
 
 ### Breaking